Modify: 댓글 동시 적용되는 것을 막기 위해, Article.js 에서 Feed를 컴포넌트화 Modify: 댓글 및 하트가 동시 적용되는 것을 막기 위해, Feed.js 에서 하트영역 컴포넌트화(LikeHeart.js) scss도 Articls.scss 에서 Feed.scss 분리
