+++ src/pages/ohj/Main/components/Nav.scss
@@ -114,7 +114,7 @@ nav {
     .btn_toLogin {
       position: absolute;
       top: 15px;
-      right: 16.5%;
+      right: 5%;
       padding: 5px 7px;
 
       border: 2px solid #222;
